ci: Use the regular build for the umc-server again
Without the cache the build time is drastically expanded because the build of
the umc-server container is quite heavy.

The task from common-ci allows to set a variable "KANIKO_CACHE_ARGS" to an empty
string in order to avoid caching, or to a different configuration as needed.
This can and should be done when manually putting a pipeline together.

Issue: #0