fix: adjust memcache containers to helm chart standards
Adjust the memcached deployment and service to our standards.
The deployment supports most of the options from the Bitnami deployment.

Everything that was previously defined in our in the umc-server
`memcached` Values.yaml section will continue to work without requiring
any changes, as well as most options from the original chart, so no
migration steps should be necessary.

Most notably, `memcached.auth.existingPasswordSecret` is not needed anymore.
Since previously `memcached.auth.existingSecret` needed to be
configured for the selfservice, this can now be used for memcached as
well.

Setting `memcached.auth.disabled=true` now will also set a stub password
for the selfservice causing it to not crash on startup.

Various settings that were not used previously are not supported.

Issue univention/dev/internal/team-nubus#1501